+++ packages/react/src/create-stencil-react-components.ts
@@ -13,6 +13,7 @@ export const createStencilReactComponents = ({
   components,
   stencilPackageName,
   customElementsDir,
+  componentsTypesDir,
   hydrateModule,
   clientModule,
   serializeShadowRoot,
@@ -21,6 +22,7 @@ export const createStencilReactComponents = ({
   components: ComponentCompilerMeta[];
   stencilPackageName: string;
   customElementsDir: string;
+  componentsTypesDir: string;
   hydrateModule?: string;
   clientModule?: string;
   serializeShadowRoot?: RenderToStringOptions['serializeShadowRoot'];
@@ -57,7 +59,7 @@ import React from 'react';
 ${createComponentImport}
 import type { EventName, StencilReactComponent } from '@stencil/react-output-target/runtime';
 ${transformTagImport}
-import type { Components } from "${stencilPackageName}/${customElementsDir}";
+import type { Components } from "${stencilPackageName}/${componentsTypesDir}";
   `
   );
 
@@ -177,6 +179,9 @@ import type { Components } from "${stencilPackageName}/${customElementsDir}";
 
     const componentEventNamesType = `${reactTagName}Events`;
 
+    const requiredProps = component.properties.filter((p) => p.required && !p.internal).map((p) => `'${p.name}'`);
+    const requiredGeneric = requiredProps.length > 0 ? `, ${requiredProps.join(' | ')}` : '';
+
     sourceFile.addTypeAlias({
       isExported: true,
       name: componentEventNamesType,
@@ -184,7 +189,7 @@ import type { Components } from "${stencilPackageName}/${customElementsDir}";
     });
 
     const transformTagParam = transformTag ? ',\n    transformTag' : '';
-    const clientComponentCall = `/*@__PURE__*/ createComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}>({
+    const clientComponentCall = `/*@__PURE__*/ createComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}${requiredGeneric}>({
     tagName: '${tagName}',
     elementClass: ${componentElement},
     // @ts-ignore - ignore potential React type mismatches between the Stencil Output Target and your project.
@@ -194,7 +199,7 @@ import type { Components } from "${stencilPackageName}/${customElementsDir}";
   })`;
 
     const getTagTransformerParam = transformTag ? ',\n    getTagTransformer' : '';
-    const serverComponentCall = `/*@__PURE__*/ createComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}>({
+    const serverComponentCall = `/*@__PURE__*/ createComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}${requiredGeneric}>({
     tagName: '${tagName}',
     properties: {${component.properties
       /**
@@ -205,7 +210,7 @@ import type { Components } from "${stencilPackageName}/${customElementsDir}";
       .map((e) => `${e.name}: '${e.attribute}'`)
       .join(',\n')}},
     hydrateModule: typeof window === 'undefined' ? (import('${hydrateModule}') as Promise<HydrateModule>) : undefined,
-    clientModule: clientComponents.${reactTagName} as StencilReactComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}>,
+    clientModule: clientComponents.${reactTagName} as StencilReactComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}${requiredGeneric}>,
     serializeShadowRoot${getTagTransformerParam}
   })`;
 
@@ -215,7 +220,7 @@ import type { Components } from "${stencilPackageName}/${customElementsDir}";
       declarations: [
         {
           name: reactTagName,
-          type: `StencilReactComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}>`,
+          type: `StencilReactComponent<${componentElement}, ${componentEventNamesType}, Components.${reactTagName}${requiredGeneric}>`,
           initializer: hydrateModule ? serverComponentCall : clientComponentCall,
         },
       ],